+++ Sourcery/Parsing/FileParser.swift
@@ -778,7 +778,7 @@ extension FileParser {
     }
 
     fileprivate func parseEnumAssociatedValues(_ body: String) -> [AssociatedValue] {
-        guard !body.isEmpty else { return [] }
+        guard !body.isEmpty else { return [AssociatedValue(localName: nil, externalName: nil, typeName: TypeName("()"))] }
 
         let items = body.commaSeparated()
         return items